Aigle/0.2/Temporal_Reasoning_Engine/Projects/raptor_0.2/requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
aiokafka==0.12.0
decord==0.6.0
docling==2.54.0
ffmpeg_python==0.2.0
PyMuPDF==1.26.4
langsmith==0.4.32
librosa==0.11.0
matplotlib==3.10.6
numpy==1.26.4
OpenCC==1.1.9
paddleocr==3.2.0
pandas==2.3.3
panns_inference==0.1.1
Pillow==11.3.0
psutil==6.1.1
python-dotenv==1.1.1
Requests==2.32.5
scipy==1.15.3
scikit-image==0.25.2
transformers==4.51.3
whisperx==3.3.1
zhconv==1.4.3
redis
platformdirs
cffi
jsonpatch
timm
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1